fix: remove embedded git repo, update IP to 168.144.33.247, update gitignore

+++ back_end/app/routes/project_routes.py
@@ -6,7 +6,7 @@ from app.service.project_service import (
     update_project,
     delete_project
     )
-from app.utils.upload import save_image
+from app.utils.upload import save_image, delete_image
 from app.utils.response import success_response, error_response
 from flask_jwt_extended import jwt_required, get_jwt_identity,get_jwt
 
@@ -36,6 +36,7 @@ def add_project():
         return error_response(message='Akses ditolak! Hanya admin yang boleh mengakses.', status_code=403)
     
     title = request.form.get('title')
+    sub_title = request.form.get('sub_title')
     description = request.form.get('description')
     demo_url = request.form.get('demo_url')
     github_url = request.form.get('github_url')
@@ -55,6 +56,7 @@ def add_project():
 
     data = {
         'title':title,
+        'sub_title': sub_title,
         'description': description,
         'demo_url': demo_url,
         'github_url': github_url,
@@ -70,8 +72,12 @@ def edit_project(id):
     claims = get_jwt() # Mengambil seluruh isi token, termasuk klaim tambahan
     if claims.get('role') != 'admin':
         return error_response(message='Akses ditolak! Hanya admin yang boleh mengakses.', status_code=403)
+    
+    old_project = get_project_by_id(id)
+    if not old_project:
+        return error_response(message='Project tidak ditemukan', status_code=404)
 
-    allowed_keys = ['title', 'description', 'demo_url', 'github_url']
+    allowed_keys = ['title', 'sub_title', 'description', 'demo_url', 'github_url']
 
     data = {key: value for key, value in request.form.items() if key in allowed_keys and value}
 
@@ -79,12 +85,14 @@ def edit_project(id):
         file = request.files['image']
         new_image_url = save_image(file)
         if new_image_url:
+            if old_project.get('image_url'):
+                delete_image(old_project['image_url'])
             data['image_url'] = new_image_url
 
-    apdated_project = update_project(id, data)
+    updated_project = update_project(id, data)
 
-    if apdated_project:
-        return success_response(data=update_project)
+    if updated_project:
+        return success_response(data=updated_project)
     
     return error_response(message='project tidak ditemukan', status_code=404)
 
@@ -95,8 +103,15 @@ def remove_project(id):
     if claims.get('role') != 'admin':
         return error_response(message='Akses ditolak! Hanya admin yang boleh mengakses.', status_code=403)
     
+    project_to_delete = get_project_by_id(id)
+    if not project_to_delete:
+        return error_response(message='Project tidak ditemukan', status_code=404)
+    old_project = project_to_delete.get('image_url')
+    
     success = delete_project(id)
 
     if success:
+        if old_project:
+            delete_image(old_project)
         return success_response(message='Project berhasil dihapus')
     return error_response(message='Project tidak ditemukan')
